BerkeleyDB实现DBManager接口完成。 明天编写TaskPool的基于BerkeleyDB的实现。

+++ src/main/java/cn/vfire/web/collector3/crawler/Fetcher.java
@@ -210,8 +210,8 @@ public class Fetcher {
 	}
 
 
-	public long getTotalCount() {
-		return (long) this.taskPool.getTotalGenerate();
+	public int getTotalCount() {
+		return this.taskPool.getTotalGenerate();
 	}
